# apps/core/openapi.yaml
openapi: 3.0.3
info:
  title: CapNet Core
  version: 0.3.0
  description: |
    CapNet Core API 초안. 실행 중 스키마는 `GET /openapi.json` · Swagger `GET /docs`.
    claim·gate_run·assignment는 INSERT…SELECT. Node는 큐를 pull하지 않는다.
    Capability 런타임 등록: POST /v1/capabilities.
servers:
  - url: http://127.0.0.1:8000
paths:
  /health:
    get:
      summary: Core·Postgres·시드 Capability 상태
      responses:
        "200":
          description: OK
  /v1/capabilities:
    get:
      summary: Capability 목록
      responses:
        "200":
          description: items[]
    post:
      summary: Capability 런타임 등록 (UNIQUE code+version, mvp CHECK)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapabilityCreate"
      responses:
        "200":
          description: 생성된 Capability
        "400":
          description: 검증·CHECK 거부
        "409":
          description: (code, version) 중복
  /v1/capabilities/{capability_id}:
    get:
      summary: Capability 단건
      parameters:
        - name: capability_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Capability
        "404":
          description: 없음
  /v1/agents:
    get:
      summary: Agent 목록
      responses:
        "200":
          description: items[]
    post:
      summary: Agent 등록 (safetensors만, .pt/.pth 거부)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreate"
      responses:
        "200":
          description: 생성된 Agent
        "400":
          description: 형식·해시 거부
  /v1/agents/{agent_id}:
    get:
      summary: Agent 단건 + capabilities
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent
        "404":
          description: 없음
  /v1/agents/{agent_id}/bindings:
    post:
      summary: Node 바인딩 → READY(해시 일치 시)
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindBody"
      responses:
        "200":
          description: binding + ready
  /v1/nodes:
    get:
      summary: Node 목록
      responses:
        "200":
          description: items[]
    post:
      summary: Node 등록 (등급은 Core가 부여, 자기주장 경로 없음)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
      responses:
        "200":
          description: 생성된 Node
  /v1/nodes/{node_id}:
    get:
      summary: Node 단건
      parameters:
        - $ref: "#/components/parameters/NodeId"
      responses:
        "200":
          description: Node
        "404":
          description: 없음
  /v1/tasks:
    post:
      summary: Task 생성 (allowlist datasetId만)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "200":
          description: Task
        "400":
          description: datasetId 거부
  /v1/tasks/{task_id}:
    get:
      summary: Task·현재 assignment·result_ref
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task
        "404":
          description: 없음
  /v1/nodes-liveness:
    get:
      summary: 기기 생존·가용 상태 (배정 근거)
      description: >
        heartbeat 신선도 · availability · 진행 중 배정 수.
        Core 워커는 살아 있고 덜 바쁜 기기에만 배정한다.
      responses:
        "200":
          description: 기기별 생존 상태 목록
  /v1/internal/nodes/{node_id}/heartbeat:
    post:
      summary: 기기 생존·가용 신고
      description: >
        AVAILABLE / BUSY / DRAINING / OFFLINE.
        신고가 끊기면(heartbeat_timeout_s 초과) 배정 후보에서 빠진다.
      parameters:
        - $ref: "#/components/parameters/NodeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                availability:
                  type: string
                  enum: [AVAILABLE, BUSY, DRAINING, OFFLINE]
                metrics:
                  type: object
                  nullable: true
      responses:
        "200":
          description: 세션 갱신됨
        "400":
          description: 알 수 없는 availability
  /v1/internal/nodes/{node_id}/assignments:
    get:
      summary: 기기가 **자기에게 배정된** 살아 있는 lease를 가져간다
      description: >
        큐 pull이 아니다. 배치는 Core 워커가 이미 끝냈고 여기서는 결과만 전달한다.
        이 경로 덕분에 Core가 기기로 들어갈 필요가 없다 (NAT).
      parameters:
        - $ref: "#/components/parameters/NodeId"
      responses:
        "200":
          description: LEASED 배정 목록 (weights_sha256 · input_ref 포함)
  /v1/internal/claim:
    post:
      summary: Core 전용 claim (FOR UPDATE SKIP LOCKED)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimBody"
      responses:
        "200":
          description: LEASED assignment 또는 빈 큐
  /v1/internal/assignments/{assignment_id}/complete:
    post:
      summary: Node 결과 보고 → SUCCEEDED/COMPLETED
      parameters:
        - $ref: "#/components/parameters/AssignmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteBody"
      responses:
        "200":
          description: 완료
        "409":
          description: lease·해시 불일치
  /v1/internal/gate-runs:
    post:
      summary: gate_run 시작 (team gate-runner만)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateStartBody"
      responses:
        "200":
          description: RUNNING
        "409":
          description: runner 아님·거부
  /v1/internal/gate-runs/{gate_run_id}:
    get:
      summary: gate_run 조회
      parameters:
        - $ref: "#/components/parameters/GateRunId"
      responses:
        "200":
          description: gate_run
        "404":
          description: 없음
  /v1/internal/gate-runs/{gate_run_id}/finish:
    post:
      summary: |
        게이트 종료. dummy=false는 실채점.
        S3 — golden_set_sha256가 gate_run 스냅샷과 불일치하면 거부.
        실게이트는 golden_set_sha256 필수.
      parameters:
        - $ref: "#/components/parameters/GateRunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateFinishBody"
      responses:
        "200":
          description: 종료·사슬 mint 여부
        "400":
          description: 지표·sha256 검증 실패
        "409":
          description: RUNNING 아님
  /openapi.yaml:
    get:
      summary: 이 YAML 정적 초안
      responses:
        "200":
          description: application/yaml
components:
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    NodeId:
      name: node_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentId:
      name: assignment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GateRunId:
      name: gate_run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    AgentCreate:
      type: object
      required: [name, version, manifest_hash, weights_uri, weights_sha256]
      properties:
        name: { type: string }
        version: { type: string }
        manifest_hash: { type: string }
        weights_uri: { type: string }
        weights_sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        weights_format: { type: string, enum: [safetensors], default: safetensors }
    NodeCreate:
      type: object
      required: [name, device_type, trust_domain, compute_tier_max]
      properties:
        name: { type: string }
        device_type: { type: string, enum: [PC_GPU, PHONE, SERVER] }
        trust_domain: { type: string, enum: [team, public] }
        compute_tier_max: { type: string, enum: [L, M, S] }
        is_gate_runner: { type: boolean, default: false }
        gpu: { type: string, nullable: true }
        provision_source: { type: string, nullable: true }
    BindBody:
      type: object
      required: [node_id, weights_sha256_seen]
      properties:
        node_id: { type: string, format: uuid }
        weights_sha256_seen: { type: string }
    TaskCreate:
      type: object
      required: [datasetId, caseId]
      properties:
        datasetId: { type: string, example: eurosat-rgb }
        caseId: { type: string }
        capability_code: { type: string, default: image.classify }
        capability_version: { type: integer, default: 1 }
        requestedAgentId: { type: string, format: uuid, nullable: true }
    ClaimBody:
      type: object
      properties:
        task_id: { type: string, format: uuid, nullable: true }
        node_id: { type: string, format: uuid, nullable: true }
    CompleteBody:
      type: object
      required: [weights_sha256, label]
      properties:
        weights_sha256: { type: string }
        label: { type: string }
        confidence: { type: number, nullable: true }
        dummy: { type: boolean, default: true }
        duration_ms: { type: integer, nullable: true }
    GateStartBody:
      type: object
      required: [agent_id, capability_id, runner_node_id]
      properties:
        agent_id: { type: string, format: uuid }
        capability_id: { type: string, format: uuid }
        runner_node_id: { type: string, format: uuid }
    GateFinishBody:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [PASSED, FAILED, ERROR] }
        golden_score: { type: number, nullable: true }
        cases_total: { type: integer, nullable: true }
        cases_passed: { type: integer, nullable: true }
        dummy: { type: boolean, default: false }
        note: { type: string, nullable: true }
        macro_f1: { type: number, nullable: true }
        invalid_rate: { type: number, nullable: true }
        min_per_class_recall:
          type: number
          nullable: true
          description: >
            계약이 min_per_class_recall 을 선언하면 실게이트 PASSED 에 필수.
            없거나 미달이면 거절된다 (라벨 공간 붕괴 차단).
        golden_set_sha256:
          type: string
          description: 실게이트 필수. gate_run 스냅샷과 불일치 시 거부 (S3).
          nullable: true
    CapabilityCreate:
      type: object
      required:
        - code
        - name
        - input_schema
        - output_schema
        - golden_set_ref
        - golden_set_sha256
        - golden_set_size
        - golden_metrics
      properties:
        code: { type: string }
        version: { type: integer, minimum: 1, default: 1 }
        name: { type: string }
        description: { type: string, nullable: true }
        input_schema: { type: object }
        output_schema: { type: object }
        output_kind:
          type: string
          enum: [closed_set_labels, structured, freeform]
          default: closed_set_labels
        compute_tier: { type: string, enum: [L, M, S], default: M }
        trust_domain_min: { type: string, enum: [team, tenant, public], default: team }
        mvp_eligible: { type: boolean, default: false }
        golden_set_ref: { type: string }
        golden_set_sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        golden_set_size: { type: integer, minimum: 1 }
        golden_metrics: { type: object }